+++ src/main/java/com/ozay/resultsetextractor/MemberResultSetExtractor.java
@@ -32,7 +32,7 @@ public class MemberResultSetExtractor implements ResultSetExtractor {
                 member.setPhone(rs.getString("phone"));
                 member.setFirstName(rs.getString("first_name"));
                 member.setLastName(rs.getString("last_name"));
-                member.setBuildingId(rs.getInt("building_id"));
+                member.setBuildingId(rs.getLong("building_id"));
                 member.setOwnership(rs.getDouble("ownership"));
                 member.setUnit(rs.getString("unit"));
                 member.setParking(rs.getString("parking"));
