+++ src/main/java/com/ozay/rowmapper/MemberRowMapper.java
@@ -19,7 +19,7 @@ public class MemberRowMapper implements RowMapper {
         member.setPhone(rs.getString("phone"));
         member.setFirstName(rs.getString("first_name"));
         member.setLastName(rs.getString("last_name"));
-        member.setBuildingId(rs.getInt("building_id"));
+        member.setBuildingId(rs.getLong("building_id"));
         member.setOwnership(rs.getDouble("ownership"));
         member.setUnit(rs.getString("unit"));
         member.setParking(rs.getString("parking"));
